refactor: move audit brief contracts to output crate
